+++ src/utils/dev/opposed-diagnostics.js
