+++ src/utils/diceHelper.js
