+++ src/utils/maps/characteristics.js
