+++ src/utils/stringHelpers.js
